requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
contourpy==1.0.5
cycler==0.11.0
flatbuffers==22.9.24
fonttools==4.37.3
idna==3.4
kiwisolver==1.4.4
loguru==0.6.0
matplotlib==3.6.0
ninja==1.10.2.4
numpy==1.23.3
onnx==1.8.1
onnx-simplifier==0.3.5
onnxoptimizer==0.3.1
onnxruntime==1.8.0
opencv-python==4.6.0.66
packaging==21.3
Pillow==9.2.0
protobuf==4.21.7
pycocotools==2.0.5
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
requests==2.28.1
six==1.16.0
tabulate==0.8.10
thop==0.1.1.post2209072238
tqdm==4.64.1
typing_extensions==4.3.0
urllib3==1.26.12
moviepy==1.0.3
pyheif==0.7.0